homework_06/views/forms/posts.py
from flask_wtf import FlaskForm
from wtforms import StringField, TextAreaField
from wtforms.validators import DataRequired, Length


class PostForm(FlaskForm):
    title = StringField("Title", validators=[DataRequired(), Length(max=256)])
    body = TextAreaField("Body", validators=[DataRequired()])
