homework_06/requirements.txt
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
# psycopg2==2.9.3
psycopg2-binary==2.9.3
